prismchatbackend/requirements.txt


fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2

langchain>=0.2.10
langchain-google-genai>=2.0.4
google-generativeai>=0.7.2

supabase>=2.5.0
httpx>=0.27.0
psycopg2-binary>=2.9

python-dotenv>=1.0.1